requested changes part 2
