+++ packages/gestalt/src/Shared/ToastSubcomponents.js
@@ -12,20 +12,24 @@ import {
 } from 'react';
 import Avatar from '../Avatar';
 import Box from '../Box';
+import ColorSchemeProvider, { useColorScheme } from '../contexts/ColorSchemeProvider';
+import { useDefaultLabelContext } from '../contexts/DefaultLabelProvider';
 import Icon from '../Icon';
 import Image from '../Image';
 import Link from '../Link';
 import Mask from '../Mask';
+import Spinner from '../Spinner';
 import Text from '../Text';
 
-const SIZE_THUMBNAIL = 40;
-const SIZE_ICON = 40;
+const SIZE_THUMBNAIL = 32;
+const SIZE_ICON = 24;
 
-export function BannerOverlayMessage({
+export function ToastMessage({
   text,
   textElement,
   helperLink,
   textColor,
+  type,
 }: {
   text: ?string | Element<'span'>,
   textElement: ?string | Element<'span'>,
@@ -39,7 +43,9 @@ export function BannerOverlayMessage({
       dangerouslyDisableOnNavigation: () => void,
     }) => void,
   },
+  type?: 'default' | 'success' | 'error' | 'progress',
 }): ReactNode {
+  const isError = type === 'error';
   const textRef = useRef<null | HTMLElement>(null);
   const [ellipsisActive, setEllipsisActive] = useState(false);
 
@@ -76,6 +82,7 @@ export function BannerOverlayMessage({
           inline
           align="start"
           color={textColor}
+          weight={isError ? 'bold' : undefined}
           lineClamp={2}
           ref={textRef}
           // Set title prop manually if text is truncated
@@ -85,7 +92,7 @@ export function BannerOverlayMessage({
           {helperLink ? (
             <Fragment>
               {' '}
-              <Text inline color={textColor}>
+              <Text inline color={textColor} weight={isError ? 'bold' : undefined}>
                 <Link
                   accessibilityLabel={helperLink.accessibilityLabel}
                   href={helperLink.href}
@@ -100,8 +107,9 @@ export function BannerOverlayMessage({
           ) : null}
         </Text>
       ) : null}
+      {/* Should the helkper link */}
       {isTruncatedWithHelperLink ? (
-        <Text color={textColor}>
+        <Text color={textColor} weight={isError ? 'bold' : undefined}>
           <Link
             accessibilityLabel={helperLink?.accessibilityLabel ?? ''}
             href={helperLink?.href ?? ''}
@@ -117,7 +125,7 @@ export function BannerOverlayMessage({
   );
 }
 
-export function BannerOverlayImageThumbnail({
+export function ToastImageThumbnail({
   thumbnail,
 }: {
   thumbnail: Element<typeof Image>,
@@ -131,18 +139,56 @@ export function BannerOverlayImageThumbnail({
   );
 }
 
-export function BannerOverlayIconThumbnail({
-  thumbnail,
-}: {
-  thumbnail: Element<typeof Icon>,
-}): ReactNode {
+export function ToastIconThumbnail({ thumbnail }: { thumbnail: Element<typeof Icon> }): ReactNode {
   return <Box aria-hidden>{cloneElement(thumbnail, { size: SIZE_ICON, color: 'inverse' })}</Box>;
 }
 
-export function BannerOverlayAvatarThumbnail({
+export function ToastAvatarThumbnail({
   thumbnail,
 }: {
   thumbnail: Element<typeof Avatar>,
 }): ReactNode {
   return <Box aria-hidden>{cloneElement(thumbnail, { size: 'sm' })}</Box>;
+}
+
+export function ToastTypeThumbnail({
+  type,
+}: {
+  type: 'default' | 'success' | 'error' | 'progress',
+}): ReactNode {
+  const { name: colorSchemeName } = useColorScheme();
+  const {
+    accessibilityIconSuccessLabel,
+    accessibilityIconErrorLabel,
+    accessibilityProcessingLabel,
+  } = useDefaultLabelContext('Toast');
+
+  return (
+    <Fragment>
+      {type === 'error' ? (
+        <Icon
+          color="inverse"
+          icon="workflow-status-problem"
+          accessibilityLabel={accessibilityIconErrorLabel}
+          size={SIZE_ICON}
+        />
+      ) : null}
+      {type === 'success' ? (
+        <ColorSchemeProvider
+          colorScheme={colorSchemeName === 'darkMode' ? 'light' : 'dark'}
+          id="icon-toast-success"
+        >
+          <Icon
+            color="success"
+            icon="workflow-status-ok"
+            accessibilityLabel={accessibilityIconSuccessLabel}
+            size={SIZE_ICON}
+          />
+        </ColorSchemeProvider>
+      ) : null}
+      {type === 'progress' ? (
+        <Spinner accessibilityLabel={accessibilityProcessingLabel} color="default" show size="sm" />
+      ) : null}
+    </Fragment>
+  );
 }
